seperate the logic into three parts and use handler
