+++ handler/handler.go
@@ -1,0 +1,24 @@
+package handler
+
+import (
+	"net/http"
+)
+
+// Hello from a snippet box
+func home(w http.ResponseWriter, r *http.Request) {
+	if r.URL.Path != "/" {
+		http.NotFound(w, r)
+		return
+	}
+	w.Write([]byte("Hello from a snippet box"))
+}
+
+func GetListingBased(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet {
+		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+		return
+	}
+	w.Header().Set("Content-Type", "application/json")
+
+	w.Write([]byte("Get a specific listing"))
+}
